pyproject.toml
[project]
name = "marrow"
version = "0.1.0"
dependencies = [
    "mss>=9.0.1",
    "pynput>=1.7.6",
    "sounddevice>=0.4.6",
    "openai-whisper>=20231117",
    "pyobjc>=10.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
]

[tool.setuptools.packages.find]
include = ["eidos*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
